Add initial codebase
Adds in the code that had been written before -- this package splits out some code from a previous project. It is still more of a gathering of scripts than a coherent package, and it is documented entirely in the form of comments in the code, in Swedish. There is no testing set up yet.

Sorting out proper documentation (in English) with Roxygen is the next step, before proceeding with refactoring and setting up testing.