Begin documentation of SCB() function
Started documenting all the parameters of the core SCB() function of the package. This documentation is still not done -- there are comments left in Swedish and parameters left without an explanation -- but there is enough here to verify that Roxygen generates documentation correctly.

Since large changes to the structure of SCB() to make it more logical and modular are planned, I won't write extensive documentation for things that are about to be changed -- that effort is better spent documenting the new, improved version of the function.