Add SCB tables
Added the actual data the package handles into sysdata.rda, so that the code has access to the data.

Chose not to expose this data to the users of the package, as the entire point of the package is to provide a unified interface to the various table formats, not to dig in the specific csv files.