BUG: Data loading & unicode in documentation
Sorted out two bugs: One with how data loading in the package works, and one where trying to use non-ASCII characters in documentation breaks Roxygen.

The fix for the latter was to just replace "Värmdö" with "Stockholm" - not an entirely satisfactory fix, but apparently this issue with unicode characters is a flaw in Roxygen and even CRAN standards, so actually fixing it properly does not appear to be feasible.