Started documentation of data
Created an xml file that documents the table about marital status. This is intended as a pattern to model documentation of the rest of the tables on.

Intention is to also create an XSD file to enable validation of these xml-formatted documentations, and write a script to automatically produce human-readable documentation of each table from these XML descriptions of the tables.

This should also make it possible to test not just the code but also the data, by testing whether the actual data conforms to the documentation in the XML file.