BUG in examples
Some of the examples contained lines of code that would be too long to fit the pdf version of the documentation. Fixed this.

Also attempted to fix some other issues where the examples wouldn't run properly. Unsure if this issue is actually fixed -- it might just be devtools::check() that's being unnecessarily harsh? The requirement to not use non-ASCII characters seems to constantly trips things up, and actually avoiding them seems very difficult.