Add more tests of data documentation
Added testing that all the columns claimed to exist do in fact exist.

There appears to be yet more problems with file encodings, since the test output contains some broken characters. But perhaps that's only because of the test package assuming all strings are ASCII -- it's not clear whether any tests are actually failing because of this, or if it just makes the output a bit uglier.