Add tests of data documentation
The code now fully checks that the things the data documentation says about the data are actually true -- the columns it says occur do occur, and have the levels the documentation claims.

The change to sysdata.rda is to include a list of municipalities for the test to check against.