Added more tests of documentation
Now started work on validating that the documentation XMLs are in the correct format. Still a little bit left to do, but it is very nearly done. Once all the tests are written, I should write a script that produces human-readable documentation of the tables from the XML documentation -- a bit uncertain of how exactly to fit that in, since it should ideally be run right before building the package? Some sort of MAKEfile?

Up next, after this part, is figuring out all the strange NOTEs from R CMD Check, and seeing if I can get one version to actually pass entirely before beginning refactoring.