More data documentation tests
Added tests that check that columns documented to be of levelsType Municipalities or NumericRange contains no extra information on levels, other than possibly max- and minimum level for a NumericRange.

Remains to add checks of columns documented to be Character, and then the entire documentation testing is finished, since all specs will be tested and data tested against all of the documentation.