Updated data documentation
Added a <title> to each data documentation file, so that the human-readable data documentation can have a nicer title.

No additional tests of this are needed -- checking that a title is present is handled by the XSD, and checking that the title is correct can't be done mechanically at all.