Finished writeTextDocumentation
The function now writes out a correctly formatted (with the right encoding -- that was the most difficult part) .Rd document documenting the XML table it was fed, as a string.

The next step is to create a suitable script to run this function on each XML file, before build time of the package.