Began implementation of a Table class
This is intended to make the SCB function able to handle data modularly, instead of having to have names of columns and levels hard-coded.

So far, what has been created is a class "Level" containing information on a single level of a column, along with tests of said class.

Next up: A class Column containing all the data about a column that one might need.