Began creating class Column
The class has been defined, along with its getters and setters, and tests of the class have been written.

What remains to do is to define the levelsDealias function, which is really the raison d'etre for the class: The possibility of passing an alias for a level of the column to the column object and automatically getting the name it has in the csv file back.

Already realising this code will need to be refactored to use inheritance instead of an internal flag, but since I don't yet know how to do inheritance in R OOP, that will have to wait until I learn how to do it. The current code works, it's just a little bit stupid.