Defined class DocumentedTable and wrote query functions
The DocumentedTable class is the kind of object that we want to be generating from our XML documentation -- essentially a rather complicated wrapper around a CSV file, containing and enforcing some constraints on it.

The query function on it is mostly just an overengineered version of subset, that can handle alternative names for columns and levels as documented in the XML. The magic will happen when we have a bunch of these DocumentedTables and have the remaining small piece of code to query all of them in a uniform way, sparing the user from having to think about which file which data point is in.

Note that all this new code is so far completely untested. So the next step is obviously to write tests of all this, and catch and fix all the hidden bugs that are inevitably there in untested code.