Refactor: Give variable a definition
In an R package, it is possible to include some data or variables in the hidden internal data of the package. This is then stored in a binary file inside the library, and at runtime, the variables can be used in the library's code.

However, this means the code refers to a variable that is never defined in the actual code -- neither its name nor what it contains. This makes the code a bit hard to understand, since normally one expect variables to be defined and given values before we use their values. In fact, since the internal data is stored in a binary format, there is no textual information on it at all, which makes for lousy version control of that variable as well.

Thus, this refactoring: Adding some code that defines the variable before it is used. There is still one variable, swe_kommuner_allpoints, that is stowed away in internal data, but it is an entire map of Sweden, so it is natively not really textual, so it is more okay to treat it as library data.