Add tests of DocumentedTable class and methods
Added a rather thorough collection of tests for the class DocumentedTable, making sure it behaves in the way we expect it to. Also fixed a couple of errors in the code defining it that were revealed by the testing.

This means the files genericTableClass.r and test-genericTableClass.r are effectively finished, or at least contain everything we need them to contain. They might still be refactored or need bug fixes, but essentially they are finished.

The next step is to write code that creates a DocumentedTable object based on the XML documentation of a table. Once that is done, we can finally rip out the old code for SCB(), which is a tangled mess of nested conditionals (a bird would love that code), and replace it with our new shiny modular algorithm.