Began files for creation of DocumentedTable objects from XML
One file contains a description of what the function will be supposed to do, and just enough fake code for it to be possible to define some tests of the code.

The tests file contains basically the tests we want to do. Unfortunately, there appears to be an issue with trying to do snapshot tests: Apparently R can't figure out how to save a DocumentedTable object to file, to keep track of its snapshot? Will have to look at that and either fix it or just remove those tests, if I'd have to actually write a save routine myself. (Since then I could as well be introducing bugs via the save functionality as via actually breaking something important... Seems not worth the effort.)