Fixed snapshot tests of createDocumentedTableFromXML
The issue turned out to be that it was not possible to automatically create a JSON save of the class objects. We can, however, instead use a less textual representation -- unfortunately, even deparse didn't work right, so we have to have these completely untextual snapshots... It's sad, but better than no tests at all.