Refactor: Remove effect-free code
Not quite dead code, since it is run, but it is code that has no actual effect -- it just set the value of a variable to the same value as what it already had.

This is certainly not a good thing to have, since it might be confusing -- a reader would normally expect code to actually do something. Since the code does nothing, it is safe to just remove it without affecting functionality.