Refactor: Replace primitive with object
Before, the data which the package queries was just stored in CSV files loaded into data.frames, which are R primitives, more or less. Now it is instead loaded into objects of a DocumentedTable class, which means they come with some documentation attached, and with better methods for querying data from them.

This commit is obviously just the tip of the iceberg that is all the code I had to write to create the documentation of the data and define the DocumentedTable class which should contain the data. But it is the step at which the old "read.csv" approach has finally been replaced.