Refactor: Substitute algorithm
The SCB() function now has a new algorithm, based on the DocumentedTable class and its query method, instead of the old algorithm, which was... Basically just a big nested conditional with a lot of things hardcoded in.

This updated algorithm also led to some other changes, as the new algorithm revealed some slight flaws in other places in the code.

I also took the chance to conform the filename of the file containing this function to the general naming scheme -- the old name was inappropriate.

Finally, the old code for the VectorSCB function was simply deleted -- it wasn't doing what it was supposed to anyway, it was only vectoring over municipalities, and otherwise returning NULL.

It remains to write tests and documentation of the SCB function -- the old partial documentation of it was removed, and new better documentation (utilising the documentation of the tables) needs to be written. It also remains to write a correct version of the vectorSCB function (as its name should be under current naming convention), and test and document it.