BUG: Incorrect checking of class of objects
The function is() returns a vector of every class an object is -- so e.g. an object of a class A inheriting  from a class B will see is() return c("A", "B") (or the other order?). So checking is(object)=="A" would give c(TRUE,FALSE), and then trying to if (is(object)=="A") gives a warning, since if() should not be fed vectors of logicals but single logicals.

The fix is to instead check whether "A" is an element of is(object). Note that this permits objects of classes inheriting from A. For us, this is not really a problem at the moment, but worth keeping in mind perhaps.