Added tests of loadAllTables
This was an underappreciated piece of the puzzle -- of course we need to be sure that we are actually successfully loading all our data into memory, or the entire package is useless.

We test that the loaded tables are of the right type and are valid objects, but do not do snapshot tests, since what tables we are loading will change as more tables are added. (And the snapshots would be exceedingly large.)

Coverage of this file: 100%