Add tests of DocumentedTable class
This is really the most important part of the entire package, since this code is the thing that does all the work of actually taking queries and finding the data we're looking for.

Implementing coverage monitoring revealed a lot of lines that were not covered, so this commit adds tests of them. It also revealed one line where a return value was of the wrong type (line 578 of DocumentedTable as of this commit), and a few lines that should be excluded from coverage monitoring since they're intentionally dead. (Final else clauses in conditionals that should cover every case -- so useful to have, if some new case is accidentally introduced, but cannot really be hit by a test, at least not without doing a lot of rather pointless work.)

Coverage: All except one line. This one line is not covered because having a setter of valueColumn doesn't play nice with the validity function. Still considering just removing the setter of valueColumn for this reason.