Tests of createDocumentedTableFromXML
The tests of this function very nearly already achieve full coverage, except for one line. This one line is the case when the specified XML is invalid.

At some point this function will be given much better checks of its XML file (essentially reusing the tests already written of the documentation), so right now there's no point in contriving a way to cover that one line that will be removed. So we just add a comment noting this fact.