Comment code. Refactor: Rename variables. Refactor: Introduce assertion
These two refactorings, along with a lot of comments explaining what is going on, are very hard to extricate from each other. The diff wouldn't make much sense if I tried to do this as three separate commits, so we get a single commit with two refactorings in it.

There are four main things that have changed, all as a consequence of actually commenting the code, making the issues obvious:
1. Many more comments.
2. fillOutMapPlotFrame was creating its own list of municipalities based on the list the map had, instead of using the same list as the rest of the package. This was changed to make sure it uses the same list as everything else.
3. A bunch of variables were renamed to conform with the naming scheme of the package.
4. A bunch of if(something){stop("Error message")}, which are basically R:ier asserts with better error messages, were added, all validating that parameters are of the right type and contain what we expect it to. Violations of these asserts would presumably have caused errors even before, but with much worse error messages, so this is indeed not changing functionality, just refactoring into a more usable structure.