Add tests of SCB() function
Added a bunch of tests of the SCB() function, which now hit nearly every line of the function. There are four lines the code coverage checker reports as not being covered, but as far as I can tell they actually are being covered? Not sure what's wrong here. The rest of the lines report as being covered.