BUG: Forgot argument to canHandleQuery, grammar in stop/warning messages
Ooops. This bug was revealed by adding tests of SCB.

Also updated some error/warning messages to make them more grammatical. Not sure if grammar errors really are a bug, maybe a bug in the programmer that wrote them. (i.e. me). But they deserve to be fixed.

Also added one comment for more readable code, but having that as a separate commit seems ridiculous.