Refactor: Introduce assertion
R doesn't really have assert() (though there is an assert function provided by a different testing library than the one I am using), so if(){stop()} is the next best thing.

What has been added is checks that the input to the fillOutMapPlotFrame has the type and shape the function expects. Presumably violations of these conditions would have caused errors even before adding this, but with stranger error messages, so this is indeed not changing functionality.