Make fillOutMapPlotFrame more easily testable
It now makes sure that the thing it returns has the standard row numbering, instead of keeping a record of how many rows it stripped out via the row numbering changing. It also is slightly more consistent in what type of NA it fills in new rows with. The latter part I am still not quite certain how it works -- it seems to matter to the tests but never matter in practice, somehow?