Added reminders of manual steps before release
When one runs devtools::release to release a new version of a package on CRAN, it runs this function to get a list of yes/no questions to ask. This serves as a reminder to perform the steps that aren't automated -- in our case, we need to run one function and do a copypaste manually.

Currently of course the package is not on CRAN, so this is just a future-proofing step, really.