Implemented vectorSCB
The primary change here is that vectorSCB.r now contains code that implements the vectorSCB function, which provides vectorisation of queries to SCB. These changes also automatically produced a documentation file, and updated DESCRIPTION and NAMESPACE with the new function and the library it needs.

There's also one line of SCB.r that had to change in order to implement vectorSCB -- specifically, how we check if forceTable was omitted had to change to make it possible to "explicitly omit" it by passing NA, since that is necessary when vectorSCB calls it using do.call.

Note that vectorSCB is so far completely untested, so adding tests of it is the next step.