Comment releaseQuestions.r, omit from codecov
The information in the comment in the file was previously only in the commit message for the file, which is less immediately accessible -- so I add it as a comment to make the code more self-contained.

Additionally, while this strictly speaking is a function, it makes no sense to test it, since it doesn't do anything. So we omit it from our code coverage calculations.