Added tests of vectorSCB function
The function is now tested. Not a lot of testing needed for this function, since it really just wraps a sapply on SCB, so most tests are run on SCB or even further down the stack.

There are two lines of vectorSCB that are not hit, and I'm not quite sure why.