Implemented vectorSCB function (#21)
* Implemented vectorSCB

The primary change here is that vectorSCB.r now contains code that implements the vectorSCB function, which provides vectorisation of queries to SCB. These changes also automatically produced a documentation file, and updated DESCRIPTION and NAMESPACE with the new function and the library it needs.

There's also one line of SCB.r that had to change in order to implement vectorSCB -- specifically, how we check if forceTable was omitted had to change to make it possible to "explicitly omit" it by passing NA, since that is necessary when vectorSCB calls it using do.call.

Note that vectorSCB is so far completely untested, so adding tests of it is the next step.

* Refactor: Introduce assertion

Again, assert() is not the R keyword, if(!thing_we_assert){stop("Informative error message")} is the R way I've learned.

This catches erroneous input, which probably isn't obviously erroneous if you don't know about R's lazy evaluation of function arguments, at an earlier state than the later exremely confusing triplicate error messages that would be thrown before if this assertion were false.

* Make example of vectorSCB more informative

The !! (bang-bang) operator is not in base R, so when recommending its use in documentation it seems like a good courtesy to tell the user which library it can be found in.

(I guess I could also make rlang load automatically when the package is loaded, but that seems a bit excessive for a thing most users probably won't need.)

[ci skip]

* Refactor: Introduce assertion

Yet more assertions making sure we throw informative errors if we are given the wrong type of argument or wrong length of argument. (The fact that a vector of logicals and a single boolean are the same type is an aspect of R I find annoying.)

* Minor: Add comments

Added a comment explaining why there is no error handling in one specific case in the function.

* Added tests of vectorSCB function

The function is now tested. Not a lot of testing needed for this function, since it really just wraps a sapply on SCB, so most tests are run on SCB or even further down the stack.

There are two lines of vectorSCB that are not hit, and I'm not quite sure why.