Refactor: Extract Subclasses
Before, there was one class Column, which had a slot "levelsType" determining which if any of the other slots of the object should be filled, and how methods should behave on the object.

Now, instead, there are three subclasses of one shared virtual class Column, one per allowed value of levelsType, each implementing methods in its own way. They also each have their own extra slots which are permitted in their case.

Note that this led to some changes to the tests of the class as well, since some of them were very tightly coupled to the particular implementation. All tests of other functions, including the tests that we still get the same output from the exported functions, pass without modification, however.

The tests of DocumentedTable still cover all except one line. This one line is uncovered essentially because of issue #23 .