Auto: Update snapshots for tests
As before, snapshots record the internal structure of classes, so they need to updated when that structure is refactored. Plus snapshots of the error messages gotten in testing validity of Dealiasable -- snapshots of errors seem like a good idea, since just expect_error might pass if we start getting an error for the wrong reason.