Enable non-numeric dataYear fields
It is sometimes nice to not need to give an integer, but instead something more descriptive, like "2018/19 school year". Thus we change the required data type here.

This required not just updating the type of the slot in the class, the class constructor, and the setter of the slot; but also the function that creates the objects from the XML files, the XSD definition of the XML files, and updating the tests to reflect these new possibilities.