De-implemented setting of valueColumn
Having this setter was problematic, see issue #23, so we are now just going to throw an error instead of even trying.

There is no functional change to the tests, just removing the comments saying to make this change, and removing one erroneous test. (Accidentally tested the wrong setter, but that test passed as well...)