Add verbose and failIfUnable parameters to vectorSCB. Refactor: Move function into class method.
The main thing that changed here is the refactoring: Before, vectorSCB went digging for itself inside the DocumentedTable objects to figure out exactly how to run its query, and then ran its queries.

Now, instead, the DocumentedTable class implements methods doing that digging and the querying, leaving a much simpler task for vectorSCB.

This refactoring does not change the functionality at all, but does make more logical sense as a structure: Questions like "can this table handle this kind of query" ought be answered by methods of the class, not by outside functions probing the object. It also makes it a lot easier to change the algorithm for actually running the vectorised queries, which is sorely needed.

As a bonus, since these changes could reuse a lot of code from SCB(), we automatically got some verbose and failIfUnable functionality from the copypasted code. Thus some revisions to the tests to handle this bonus behaviour, and an update to the documentation to document it.