Add getColumn parameter to columnDealias method
The old, default, behaviour of columnDealias is to return just the real name of the column. But sometimes, we do this because we then want to dealias some levels of the column.

Before, we'd then have to loop through the list of columns to figure out which has the name we dealiased to, to then run levelDealias against that one. Now we just get the Column object directly by setting getColumn=TRUE, which makes for a faster and less confusing algorithm.