Refactor: Substitute algorithm in vectorSCB
Our old algorithm was effectively the naive algorithm, just using a sapply: This meant that we need to loop through the entire raw data once per level we are vectoring over. When, as in our example, wanting to count the number of men in each municipality, this meant we had to loop over all 234k rows 290 times.

The new and improved algorithm is cleverer, and only needs to look at each entry of our 234k-row-5-column dataset exactly once, since it makes use of the fact that it knows we're trying to vectorise a query.

This change means the specific example has gone from running in about 80 seconds to 1.2 seconds. A vast improvement.